WIP: Remove Dashboard and move /delete tests
